Se agregan dataSources y repository para movieDetail. Se agrega vista a grafo de navgacion. Se agrega paso de argumentos para consultal el detalle de una pelicula. Falta crear y dibujar la vista.

+++ app/src/main/java/com/arthur/tmdb/data/remote/api/MovieApi.kt
@@ -12,7 +12,7 @@ interface MovieApi {
 
     @GET("movie/{movieId}")
     suspend fun getMovie(
-        @Path("movieId") id: Int
+        @Path("movieId") id: Long
     ): Response<MovieDetailDto>
 
     @GET("movie/now_playing")
