+++ app/src/main/java/com/arthur/tmdb/data/remote/dto/MovieDetailDto.kt
@@ -9,7 +9,7 @@ data class MovieDetailDto(
     @SerializedName("budget") val budget: Int?,
     @SerializedName("genres") val genres: List<Genres>?,
     @SerializedName("homepage") val homepage: String?,
-    @SerializedName("id") val id: Int?,
+    @SerializedName("id") val id: Long?,
     @SerializedName("imdb_id") val imdbId: String?,
     @SerializedName("original_language") val originalLanguage: String?,
     @SerializedName("original_title") val originalTitle: String?,
@@ -26,7 +26,7 @@ data class MovieDetailDto(
     @SerializedName("tagline") val tagline: String?,
     @SerializedName("title") val title: String?,
     @SerializedName("video") val video: Boolean?,
-    @SerializedName("vote_average") val voteAverage: Double?,
+    @SerializedName("vote_average") val voteAverage: Float?,
     @SerializedName("vote_count") val voteCount: Int?
 )
